move list of internal tool paths to const file
