+++ airbyte-ci/connectors/pipelines/pipelines/dagger/containers/internal_tools.py
@@ -5,7 +5,7 @@
 from dagger import Container, Secret
 from pipelines.dagger.actions.python.pipx import with_installed_pipx_package
 from pipelines.dagger.containers.python import with_python_base
-from pipelines.internal_tools.internal import INTERNAL_TOOL_PATHS
+from pipelines.consts import INTERNAL_TOOL_PATHS
 from pipelines.pipeline.connectors.context import PipelineContext
