--- airbyte-ci/connectors/pipelines/pipelines/internal_tools/__init__.py
@@ -1,3 +1,0 @@
-#
-# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
-#
