+++ airbyte-ci/connectors/pipelines/pipelines/pipeline/metadata/pipeline.py
@@ -11,7 +11,7 @@ from pipelines.dagger.actions.python.common import with_pip_packages
 from pipelines.dagger.containers.python import with_python_base
 from pipelines.helpers.steps import run_steps
 from pipelines.helpers.utils import DAGGER_CONFIG, get_secret_host_variable
-from pipelines.internal_tools.internal import INTERNAL_TOOL_PATHS
+from pipelines.consts import INTERNAL_TOOL_PATHS
 from pipelines.models.reports import Report
 from pipelines.models.steps import MountPath, Step, StepResult
 from pipelines.pipeline.connectors.context import ConnectorContext, PipelineContext
